+++ src/Draco.Lsp/Model/NotificationMessage.cs
@@ -1,18 +1,13 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using Draco.Lsp.Model;
 
-namespace Draco.Lsp.Serialization;
+namespace Draco.Lsp.Model;
 
-internal sealed class RequestMessage
+internal sealed class NotificationMessage
 {
     [JsonPropertyName("jsonrpc")]
     [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
     public required string Jsonrpc { get; set; }
-
-    [JsonPropertyName("id")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public required OneOf<int, string> Id { get; set; }
 
     [JsonPropertyName("method")]
     [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
